test: have DataGateway API and search API separated tests
