+++ test/datagateway_api/icat/endpoints/test_ping_icat.py
