+++ test/datagateway_api/icat/endpoints/test_update_multiple_icat.py
@@ -1,6 +1,6 @@
 import pytest
 
-from test.icat.test_query import prepare_icat_data_for_assertion
+from test.datagateway_api.icat.test_query import prepare_icat_data_for_assertion
 
 
 class TestUpdateMultipleEntities:
