+++ test/datagateway_api/icat/test_session_handling.py
